feat: add dataloader for celeba
