+++ phydir/datasets/dataloaders.py
@@ -4,6 +4,8 @@ import torchvision.transforms as tfs
 import torch.utils.data
 import numpy as np
 from PIL import Image
+from torch.utils.data import ConcatDataset
+from common import ImageDataset
 
 
 def get_data_loaders(cfgs):
@@ -12,18 +14,15 @@ def get_data_loaders(cfgs):
     image_size = cfgs.get('image_size', 256)
     crop = cfgs.get('crop', None)
 
+    training_data = cfgs.get('training_data', 'celeba')
     run_train = cfgs.get('run_train', False)
     train_val_data_dir = cfgs.get('train_val_data_dir', './data')
     run_test = cfgs.get('run_test', False)
     test_data_dir = cfgs.get('test_data_dir', './data/test')
 
-    load_gt_depth = cfgs.get('load_gt_depth', False)
-    AB_dnames = cfgs.get('paired_data_dir_names', ['A', 'B'])
-    AB_fnames = cfgs.get('paired_data_filename_diff', None)
-
     train_loader = val_loader = test_loader = None
 
-    get_loader = lambda **kargs: get_image_loader(**kargs, batch_size=batch_size, image_size=image_size, crop=crop)
+    get_loader = lambda **kargs: get_image_loader(**kargs, datasets=training_data, batch_size=batch_size, image_size=image_size, crop=crop)
 
     if run_train:
         train_data_dir = os.path.join(train_val_data_dir, "train")
@@ -41,70 +40,17 @@ def get_data_loaders(cfgs):
 
     return train_loader, val_loader, test_loader
 
-
-IMG_EXTENSIONS = ('.jpg', '.jpeg', '.png', '.ppm', '.bmp', '.pgm', '.tif', '.tiff', 'webp')
-def is_image_file(filename):
-    return filename.lower().endswith(IMG_EXTENSIONS)
-
-
-## simple image dataset ##
-def make_dataset(dir):
-    dir = '../data/img_celeba'
-    assert os.path.isdir(dir), '%s is not a valid directory' % dir
-    images = {}
-    for root, _, fnames in sorted(os.walk(os.path.join(dir, 'datalist'))):
-        for fname in fnames:
-            dlist = json.load(open(os.path.join(root, fname)))
-            if len(dlist) >= 6:
-                n_id, _ = fname.split('.')
-                images[n_id] = dlist
-    return images
-
-
-class ImageDataset(torch.utils.data.Dataset):
-    def __init__(self, data_dir, image_size=256, crop=None, is_validation=False):
-        super(ImageDataset, self).__init__()
-        self.root = data_dir
-        self.image_size = image_size
-        self.ids = make_dataset(data_dir)
-        self.size = len(self.ids) # len(self.paths.keys()) # todo check size
-        self.crop = crop
-        self.is_validation = is_validation
-
-    def transform(self, img, hflip=False):
-        if self.crop is not None:
-            if isinstance(self.crop, int):
-                img = tfs.CenterCrop(self.crop)(img)
-            else:
-                assert len(self.crop) == 4, 'Crop size must be an integer for center crop, or a list of 4 integers (y0,x0,h,w)'
-                img = tfs.functional.crop(img, *self.crop)
-        img = tfs.functional.resize(img, (self.image_size, self.image_size))
-        if hflip:
-            img = tfs.functional.hflip(img)
-        return tfs.functional.to_tensor(img)
-
-    def __getitem__(self, index):
-        K = np.random.randint(1, 7) # random 1~6
-        random_ind = np.random.permutation(len(self.ids[index]))[:K]
-        imgs = []
-        for fpath in random_ind:
-            if is_image_file(fpath):
-                img = Image.open(fpath).convert('RGB')
-                hflip = not self.is_validation and np.random.rand() > 0.5
-                img = self.transform(img, hflip)
-                imgs.append(img)
-        imgs = torch.stack(imgs, dim=0) # [K, 3, 256, 256]
-        return imgs
-
-    def __len__(self):
-        return self.size
-
-    def name(self):
-        return 'ImageDataset'
-
-
-def get_image_loader(data_dir, is_validation=False,
+def get_image_loader(data_dir, is_validation=False, datasets='celeba',
     batch_size=8, num_workers=4, image_size=256, crop=None):
+    # data_list = []
+
+    # todo
+    if 'celeba' in datasets:
+        pass
+    # if 'casia' in cfg.trainig_data:
+        # data_list.append()
+        # pass
+    # dataset = ConcatDataset(data_list)
 
     dataset = ImageDataset(data_dir, image_size=image_size, crop=crop, is_validation=is_validation)
     loader = torch.utils.data.DataLoader(
@@ -115,7 +61,6 @@ def get_image_loader(data_dir, is_validation=False,
         pin_memory=True
     )
     return loader
-
 
 if __name__ == '__main__':
     import argparse
@@ -148,4 +93,3 @@ if __name__ == '__main__':
         print(data.size())
         if i > 0:
             break
-
